Setup command line argument parsing
For command line parsing 'argparse' is used instead of tf.flags. The rational behind the choice is that it is more generic than the tensorflow solution and has, to my best knowledge, more features. For example checking for allowed values for each argument is already done by argparse which is quite useful.