Setup class structure, dependencies and interfaces
The overall design idea is, that all implementation goes into the 'algos' folder. 
Only the interfaces have to be fullfilled, the implementation of the reinforcement algorithm can be as desired. 
The interfaces are defined by abstract classes for the Manager, Trainer and Runner. The Manager is the central entry point. Basic descision e.g. wether to train or evaluate are made at this level. Also the thread for Runners and Trainers are started here.
There are TrainerAgents and RunnerAgents. The RunnerAgents only collect experiences while the TrainerAgent uses these, guided by the Trainer, to learn.
This approach allows for parallization and follows the design idea of the A3C algorithm. 

Arguments are handed down from the Manager to Trainer/Runners to Agents. 
class:    arguments
M:    m, a, g, tb, cp, sw
R,T:        a, g, tb, cp, sw, q
TA:                 tb, cp, sw, bn
RA:                 tb, cp, sw, bn, ta