Setup basic threading and queue sharing for 'a3c' algorithm
The build-in modules 'threading' and 'queue' are used instead of the tensorflow module because the tensorflow queue is not yet supported in eager mode.

For this implementation of the 'a3c' algorithm all threads are stopped once the trainer thread has been running 'num_episode' times. This mights change in the future since it might lead to inefficient usage of the experiences of the runners (when runners are much faster than the trainer) .