Setup run method for multi environment support
It is now possible to run different OpenAI gym environment. Training is not yet implemented. 
The environments are different w.r.t. the their action space (continuous or discrete) and the observation space (frame based of quasi state space). Depending of the given game environment different nets are used and the action sampling is adapted. In case of discrete action spaces sampling from the policy distribution is done via argmax. In case of continuous action spaces sampling is done according to a multi-gaussian distribution.